Force doxygen generation to be dependent on source file presence.
The dependencies start with Doxyfile which is produced by configure, so
thereafter no updates to these documents will happen by make alone.
Re-run configure and make, or, delete the stamp file and run make to
force an update.

Test for the -lib and -dev sources to ensure that the Doxygen
documentation and resulting tarball are updated only when the sources
are present. The importance of this is that during make install for this
-doc package the sources are not present and make all would would
produce basic empty Doxygen documentation which would then be installed.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>