lib: buf-msg: get_msg to process partial messages without re-processing.
get_msg now accommodates partial messages without re-processing the
entire buffer. Data from the buffer extracted to the message struct
could be a complete message, (terminated with the
mgemessage.terminator), or a partial message. If it is an incomplete
message then data will be appended to the partial message next time this
function is invoked, repeating until a complete message is held in the
struct.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>