buf-msg: Ensure partially-processed buffer support
Entry to get_msg() assumed a starting offset of zero in the supplied
buffer object. This worked because trim-buf() was always called
subsequently which reset proc_next to 0. However it should cope with
partially-processed buffers where trim_buf() has not been called and the
starting offset proc_next is not zero.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>