buf-msg: Move trim_buf() from get_msg to pull_msg
Trimming the buffer object may be a logical thing to do after getting a
message but it is not mandatory and the two are not conceptually
hard-linked. So, move trim_buf() out of get_msg() and into pull_msg(),
after the get_msg(), to keep the implementation as flexible as the
design.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>