buf-msg: trim_buf() improve algorithm
Re-sizing the buffer every time this function is called would not be
efficient. So, the function first tests whether the buffer has a defined
multiple of the default buffer size free, and if so shrinks the buffer and
relocates valid data to the start of the buffer. If that test does not
result in action, the second test determines if the percentage of
unreachable portion of the buffer to the size of the buffer exceeds a
defined value, and if so, it relocates data to the start of the buffer.

Signed-off-by: Mark Grant <m.grant.prg@gmail.com>