docs: Doxygen silently fails to incorporate dia files
Doxygen output does not contain the dia images specified with the
Doxygen directive @diafile. This is beacuse Doxygen hard codes the
use of the png-libart export filter which is not maintained and is no
longer available on many distros. In this case Doxygen / dia silently
fails and no drawings are generated or included.

The workaround is to manually generate png files from the dia files
using the png export filter and include them in the generated
documentation using the @image html filename directive.

Link: https://github.com/doxygen/doxygen/blob/d4beb392b452d14d375e1406ec4fb0f6436081ee/src/dia.cpp#L46
Signed-off-by: Mark Grant <m.grant.prg@gmail.com>